refactor(capabilities): split the audio cap into identity and runtime modules
Drop `#[bridge]` + `mod native` and move the `aether.audio` cap onto the
ADR-0122 identity/runtime split: a ZST `AudioCapability` identity plus an
`AudioCapabilityState` in a new `audio/runtime.rs`, gated on the media cap's
own `audio-native` feature via `#[actor(singleton, runtime_feature =
"audio-native")]`. Handlers become associated fns over `state: &mut
Self::State`. The cpal worker-thread teardown follows the `thread` +
`shutdown` fields onto the state via `impl Drop for AudioCapabilityState`,
so the dispatcher's drop stops and joins the worker without a dedicated
`unwire`. Behaviour-preserving: identity, kind ids, NAMESPACE, and the
`audio-native` feature are unchanged.

Closes #2332
